new Metal filter to simulate? dog eyes
Dogs are dichromats. They can "see" yellow & blue, but can't see red.
To simulate dog's dichromat eyes, I set the green component of the pixel
to the same value as red -- because green + red = yellow. The blue component
and alpha values are unchanged.

This is probably the most naive "dog vision" ever implemented